backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.32.0
sqlmodel==0.0.24
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.5
anthropic==0.7.7
pydantic[email]==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
pillow==10.1.0
librosa==0.10.1
soundfile==0.12.1
python-socketio==5.10.0
websockets==12.0
celery==5.3.4
flower==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
